requirements.txt
Django==4.2.10
django-ninja==1.0.1
django-environ==0.11.2
celery==5.3.6
redis==5.0.1
mysqlclient==2.2.1
gunicorn==21.2.0
pyjwt==2.8.0
pydantic==2.6.1
pytest==7.4.4
pytest-django==4.7.0
python-dateutil==2.8.2
openpyxl==3.1.2
Pillow==10.1.0 